+++ app.vue
@@ -2,9 +2,17 @@
   <div>
     <NuxtPage />
   </div>
+  <FooterApiFooter />
 </template>
 
 <script lang="ts" setup>
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { faUser } from '@fortawesome/free-regular-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+
+library.add(faUser)
+
+
 
 useHead(() => ({
   link: [