+++ services/apiService.ts
@@ -1,9 +1,11 @@
 import axios from 'axios';
 
-export const getRandomFirstName = async () => {
+export const getRandomCharacter = async () => {
   try {
-    const response = await axios.get('http://localhost:3001/pnj/random-firstname');
-    return response.data.firstname;
+    const response = await axios.get('http://localhost:3001/apocalypse/random-pnj');
+    console.log('response.data', response.data);
+    return response.data;
+    
   } catch (error) {
     console.error('Erreur lors de l\'appel à l\'API :', error);
     return null;